#17737: move matmul sd tests to nightly and adjust matmul test dimensions
